Rename ui template css service
